♻️ refactor: Move BaldDataset static methods to data_utils as standalone functions
