+++ .pre-commit-config.yaml
@@ -1,12 +1,12 @@
 repos:
+  - repo: https://github.com/pre-commit/mirrors-isort
+    rev: "v5.10.1"
+    hooks:
+      - id: isort
+        args: ["--profile", "black", "--line-length", "79"]
+
   - repo: https://github.com/psf/black
     rev: 23.1.0
     hooks:
       - id: black
         args: [--line-length=79]
-
-  - repo: https://github.com/pre-commit/mirrors-isort
-    rev: "v5.10.1"
-    hooks:
-      - id: isort
-        args: ["--profile", "black"]
