+++ scripts/prepare_datasets.py
@@ -1,8 +1,17 @@
 import os
+
 import pandas as pd
+
+from data_utils import (
+    convert_image_id_column_to_float,
+    create_subset_dfs,
+    get_cleaned_df,
+    prepare_merged_dataframe,
+    replace_bald_label,
+)
 from src.config_class import BaldOrNotConfig
 from src.constants import NOT_BALD_LABEL, ORIGINAL_NOT_BALD_LABEL
-from src.data import BaldDataset
+from src.dataset import BaldDataset
 
 
 def main():
@@ -16,19 +25,17 @@ def main():
     subsets_df = pd.read_csv(subsets_division_ds_path)
     labels_df = pd.read_csv(labels_ds_path)
 
-    cleaned_df = BaldDataset.get_cleaned_df(labels_df, images_dir)
-    merged_df = BaldDataset.prepare_merged_dataframe(subsets_df, labels_df)
-    converted_df = BaldDataset.convert_image_id_column_to_float(merged_df)
-    corrected_labels_df = BaldDataset.replace_bald_label(
+    cleaned_df = get_cleaned_df(labels_df, images_dir)
+    merged_df = prepare_merged_dataframe(subsets_df, cleaned_df)
+    converted_df = convert_image_id_column_to_float(merged_df)
+    corrected_labels_df = replace_bald_label(
         converted_df,
         original_label=ORIGINAL_NOT_BALD_LABEL,
         new_label=NOT_BALD_LABEL,
     )
 
     # Create data subsets
-    train_df, val_df, test_df = BaldDataset.create_subset_dfs(
-        corrected_labels_df
-    )
+    train_df, val_df, test_df = create_subset_dfs(corrected_labels_df)
 
     # Save subsets to CSV files
     data_dir = os.path.join("..", "src", "data")
