+++ tests/test_dataset.py
@@ -4,19 +4,24 @@ import numpy as np
 import pandas as pd
 import pytest
 
+from data_utils import (
+    create_subset_dfs,
+    get_cleaned_df,
+    get_wrong_files_list,
+    prepare_merged_dataframe,
+    undersample_classes,
+)
 from src.constants import (
     DEFAULT_IMG_SIZE,
     N_CHANNELS_GRAYSCALE,
     N_CHANNELS_RGB,
 )
-from src.data import BaldDataset
+from src.dataset import BaldDataset
 
 
 @pytest.fixture
 def sample_df():
-    """
-    Fixture that provides a sample DataFrame to use for testing.
-    """
+    """Fixture that provides a sample DataFrame to use for testing."""
     data = {
         "image_id": ["bald.jpg", "not_bald.jpg"],
         "labels": [1, 0],
@@ -27,17 +32,12 @@ def sample_df():
 
 @pytest.fixture
 def dataset(sample_df):
-    """
-    Fixture that initializes the BaldDataset class with the sample DataFrame.
-    """
+    """Fixture that initializes the BaldDataset class with the sample DataFrame."""
     return BaldDataset(sample_df, batch_size=1, shuffle=False)
 
 
 def test_init(sample_df):
-    """
-    Test the initialization of the BaldDataset class to ensure default
-    parameters are set correctly.
-    """
+    """Test the initialization of the BaldDataset class to ensure default parameters are set correctly."""
     dataset = BaldDataset(sample_df)
     assert dataset.batch_size == 32
     assert dataset.dim == DEFAULT_IMG_SIZE
@@ -50,21 +50,16 @@ def test_init(sample_df):
     "n_channels, expected", [(N_CHANNELS_RGB, 3), (N_CHANNELS_GRAYSCALE, 1)]
 )
 def test_n_channels_accepts_valid_values(sample_df, n_channels, expected):
-    """
-    Test that valid values for n_channels are accepted by the
-    BaldDataset class.
-    """
+    """Test that valid values for n_channels are accepted by the BaldDataset class."""
     dataset = BaldDataset(sample_df, n_channels=n_channels)
     assert dataset.n_channels == expected, f"n_channels should be {expected}."
 
 
 def test_n_channels_rejects_invalid_value(sample_df):
-    """
-    Test that invalid values for n_channels raise a ValueError.
-    """
+    """Test that invalid values for n_channels raise a ValueError."""
     with pytest.raises(
         ValueError,
-        match="n_channels must be either 1 \(grayscale\) or 3 \(RGB\)\.",  # noqa: W605, E501
+        match="n_channels must be either 1 \\(grayscale\\) or 3 \\(RGB\\)\\.",
     ):
         BaldDataset(sample_df, n_channels=2)
 
@@ -77,20 +72,16 @@ def test_n_channels_rejects_invalid_value(sample_df):
     ],
 )
 def test_len(sample_df, batch_size, expected_length):
-    """
-    Test that the __len__ method returns the correct number of batches
-    based on batch_size.
-    """
+    """Test that the __len__ method returns the correct number of batches
+    based on batch_size."""
     dataset = BaldDataset(sample_df, batch_size=batch_size)
     assert len(dataset) == expected_length
 
 
 @pytest.mark.parametrize("shuffle", [True, False])
 def test_on_epoch_end(sample_df, shuffle):
-    """
-    Test the on_epoch_end method to ensure indexes are shuffled correctly
-    when shuffle is True, and remain the same when shuffle is False.
-    """
+    """Test the on_epoch_end method to ensure indexes are shuffled correctly
+    when shuffle is True, and remain the same when shuffle is False."""
     dataset = BaldDataset(sample_df, shuffle=shuffle)
     initial_indexes = dataset.indexes.copy()
     dataset.on_epoch_end()
@@ -103,10 +94,7 @@ def test_on_epoch_end(sample_df, shuffle):
 
 
 def test_getitem_calculates_indices_correctly(dataset):
-    """
-    Test that the __getitem__ method correctly calculates indices for
-    each batch.
-    """
+    """Test that the __getitem__ method correctly calculates indices for each batch."""
     # Test the first batch
     expected_indices = [0]  # Indices for the first batch
     actual_indices = dataset.indexes[0:1]
@@ -123,10 +111,7 @@ def test_getitem_calculates_indices_correctly(dataset):
 
 
 def test_getitem_extracts_image_ids_correctly(dataset):
-    """
-    Test that the __getitem__ method extracts the correct image IDs for
-    each batch.
-    """
+    """Test that the __getitem__ method extracts the correct image IDs for each batch."""
     # Test the first batch
     expected_list_IDs_temp = ["bald.jpg"]
     actual_list_IDs_temp = [dataset.list_IDs[i] for i in dataset.indexes[0:1]]
@@ -146,10 +131,7 @@ def test_getitem_extracts_image_ids_correctly(dataset):
 def test_getitem_calls_data_preprocessing_correctly(
     mock_data_preprocessing, sample_df
 ):
-    """
-    Test that the __getitem__ method calls the __data_preprocessing method
-    correctly with the right image IDs.
-    """
+    """Test that the __getitem__ method calls the __data_preprocessing method correctly with the right image IDs."""
     mock_data_preprocessing.return_value = (
         np.zeros((2, *DEFAULT_IMG_SIZE, 3)),  # Mocked X (images)
         np.array([1, 0]),  # Mocked y (labels)
@@ -162,75 +144,28 @@ def test_getitem_calls_data_preprocessing_correctly(
     expected_list_IDs_temp = ["bald.jpg", "not_bald.jpg"]
     mock_data_preprocessing.assert_called_with(expected_list_IDs_temp)
 
-    # There is no second batch, so no need to simulate it
-
-
-@patch.object(BaldDataset, "_BaldDataset__data_preprocessing")
-def test_getitem_returns_correct_X_and_y(mock_data_preprocessing, dataset):
-    """
-    Test that the __getitem__ method returns the correct X (images) and
-    y (labels) for each batch.
-    """
-    # Mock the return value of __data_preprocessing
-    mock_data_preprocessing.return_value = (
-        np.array([[[[0.1]], [[0.2]], [[0.3]]]]),  # Mocked X (images)
-        np.array([1, 0]),  # Mocked y (labels)
-    )
-
-    # Test the first (and only) batch
-    X, y = dataset[0]
-    expected_X = np.array([[[[0.1]], [[0.2]], [[0.3]]]])
-    expected_y = np.array([1, 0])
-    assert np.array_equal(
-        X, expected_X
-    ), "Returned X (images) for the first batch is incorrect."
-    assert np.array_equal(
-        y, expected_y
-    ), "Returned y (labels) for the first batch is incorrect."
-
-    # There is no second batch, so no need to test it
-
 
 def test_get_wrong_files_list():
-    """
-    Test the __get_wrong_files_list method to ensure it correctly identifies
-    files that cannot be read as images.
-    """
+    """Test the get_wrong_files_list function to ensure it correctly identifies files that cannot be read as images."""
     with patch("os.listdir", return_value=["bald.jpg", "not_bald.jpg"]):
-        with patch(
-            "cv2.imread", side_effect=[None, np.ones((300, 300, 3))]
-        ):  # noqa: E501
-            wrong_files = BaldDataset._BaldDataset__get_wrong_files_list(
-                "src/samples"
-            )
+        with patch("cv2.imread", side_effect=[None, np.ones((300, 300, 3))]):
+            wrong_files = get_wrong_files_list("src/samples")
             assert len(wrong_files) == 1
             assert wrong_files[0] == "bald.jpg"
 
 
 def test_get_cleaned_df(sample_df):
-    """
-    Test the get_cleaned_df method to ensure it correctly removes rows
-    corresponding to images that cannot be read.
-    """
+    """Test the get_cleaned_df function to ensure it correctly removes rows corresponding to images that cannot be read."""
     with patch(
-        "src.data.BaldDataset._BaldDataset__get_wrong_files_list",
-        return_value=["not_bald.jpg"],
+        "data_utils.get_wrong_files_list", return_value=["not_bald.jpg"]
     ):
-        cleaned_df = BaldDataset.get_cleaned_df(
-            sample_df, images_dir="src/samples"
-        )
+        cleaned_df = get_cleaned_df(sample_df, images_dir="src/samples")
         assert len(cleaned_df) == 1
         assert "not_bald.jpg" not in cleaned_df["image_id"].values
 
 
 def test_prepare_merged_dataframe(tmpdir):
-    """
-    Test the prepare_merged_dataframe method to ensure it merges two CSV files
-    into a single DataFrame with the correct columns.
-    """
-    subsets_path = tmpdir.join("subsets.csv")
-    labels_path = tmpdir.join("labels.csv")
-
+    """Test the prepare_merged_dataframe function to ensure it merges two CSV files into a single DataFrame with the correct columns."""
     subsets_df = pd.DataFrame(
         {"image_id": ["bald.jpg", "not_bald.jpg"], "partition": [0, 1]}
     )
@@ -238,26 +173,21 @@ def test_prepare_merged_dataframe(tmpdir):
         {"image_id": ["bald.jpg", "not_bald.jpg"], "Bald": [1, 0]}
     )
 
-    subsets_df.to_csv(subsets_path, index=False)
-    labels_df.to_csv(labels_path, index=False)
-
-    df_merged = BaldDataset.prepare_merged_dataframe(subsets_path, labels_path)
+    df_merged = prepare_merged_dataframe(subsets_df, labels_df)
 
     assert isinstance(
         df_merged, pd.DataFrame
     ), "The result should be a DataFrame."
-    assert list(df_merged.columns) == ["image_id", "partition", "labels"], (
-        "The DataFrame should contain only the columns: 'image_id', "
-        "'partition', and 'labels'."
-    )
+    assert list(df_merged.columns) == [
+        "image_id",
+        "partition",
+        "label",
+    ], "The DataFrame should contain only the columns: 'image_id', 'partition', and 'label'."
     assert len(df_merged) == 2, "The DataFrame should contain 2 rows."
 
 
 def test_create_subset_dfs(sample_df):
-    """
-    Test the create_subset_dfs method to ensure it correctly creates
-    training, validation, and test subsets from the DataFrame.
-    """
+    """Test the create_subset_dfs function to ensure it correctly creates training, validation, and test subsets from the DataFrame."""
     extra_data = pd.DataFrame(
         {
             "image_id": ["new_sample.jpg"],
@@ -268,60 +198,45 @@ def test_create_subset_dfs(sample_df):
 
     extended_sample_df = pd.concat([sample_df, extra_data], ignore_index=True)
 
-    train_df, val_df, test_df = BaldDataset.create_subset_dfs(
-        extended_sample_df
-    )
+    train_df, val_df, test_df = create_subset_dfs(extended_sample_df)
     assert len(train_df) == 1
     assert len(val_df) == 1
     assert len(test_df) == 1
     assert "partition" not in test_df.columns
 
 
-def test_undersample_majority_class():
+def test_undersample_classes():
+    """Test the undersample_classes function to ensure it correctly undersamples the majority class based on the provided ratio."""
     data_majority = {
         "feature1": range(9000),
         "feature2": range(9000, 18000),
-        "label": [0] * 9000,  # Majority class label is 0
+        "label": [0] * 9000,
     }
-
     data_minority = {
         "feature1": range(1000),
         "feature2": range(18000, 19000),
-        "label": [1] * 1000,  # Minority class label is 1
+        "label": [1] * 1000,
     }
 
     df_majority = pd.DataFrame(data_majority)
     df_minority = pd.DataFrame(data_minority)
 
     df_unbalanced = pd.concat([df_majority, df_minority], ignore_index=True)
-
     df_unbalanced = df_unbalanced.sample(frac=1, random_state=42).reset_index(
         drop=True
     )
-    desired_ratio = 2.0
-    df_balanced = BaldDataset.undersample_majority_class(
-        df_unbalanced,
-        label_col="label",
-        majority_class_label=0,
-        desired_classes_ratio=desired_ratio,
+
+    class_sample_sizes = {0: 2000, 1: 1000}
+    df_balanced = undersample_classes(
+        df_unbalanced, label_col="label", class_sample_sizes=class_sample_sizes
     )
 
     n_minority = len(df_balanced[df_balanced["label"] == 1])
-    expected_n_majority = int(n_minority * desired_ratio)
-
     n_majority = len(df_balanced[df_balanced["label"] == 0])
 
     assert (
-        n_majority == expected_n_majority
-    ), f"Expected {expected_n_majority} majority samples, got {n_majority}"
-
-    expected_total_samples = n_minority + expected_n_majority
-    actual_total_samples = len(df_balanced)
+        n_majority == class_sample_sizes[0]
+    ), f"Expected {class_sample_sizes[0]} majority samples, got {n_majority}"
     assert (
-        actual_total_samples == expected_total_samples
-    ), f"Expected {expected_total_samples} samples, got {actual_total_samples}"
-
-    original_n_minority = len(df_unbalanced[df_unbalanced["label"] == 1])
-    assert (
-        n_minority == original_n_minority
-    ), "There should be no change in minority samples"
+        n_minority == class_sample_sizes[1]
+    ), f"Expected {class_sample_sizes[1]} minority samples, got {n_minority}"
